Changed behaviour of any action functions
Previously keys that existed in the deck but didn't have any action
functions would be default be trapped by the "any" functions if any
were specified. This behaviour is redudant because you have to trap
their string in the any function anyway. "Blank" keys that are
recognized but trigger no actions are more useful